implemented first simple relationship and some test services

+++ shero/src/main/java/edu/upc/ase/service/ItemService.java
@@ -1,7 +1,5 @@
 package edu.upc.ase.service;
 
-import java.math.BigDecimal;
-
 import edu.upc.ase.domain.Item;
 
 public class ItemService {
@@ -9,7 +7,7 @@ public class ItemService {
 	
 	public Item getExampleItem() {
 		// should retrieve item from DB in future, based on some id for example
-		Item item = new Item("Surfboard XYZ", new BigDecimal(27.50));
+		Item item = new Item("Surfboard XYZ", new Double(27.50));
 		return item;
 	}
 }
